+++ internal/external-dns/registry/name_mapper_externaldns.go
@@ -2,19 +2,7 @@ package registry
 
 import (
 	"strings"
-
-	"github.com/kuadrant/dns-operator/internal/common/hash"
 )
-
-/**
-  nameMapper is the interface for mapping between the endpoint for the source
-  and the endpoint for the TXT record.
-*/
-
-type nameMapper interface {
-	toEndpointName(txtDNSName, version string) (endpointName, recordType string)
-	toTXTName(string, string, string) string
-}
 
 type externalDNSAffixNameMapper struct {
 	prefix              string
@@ -149,77 +137,4 @@ func (pr externalDNSAffixNameMapper) toTXTName(endpointDNSName, _, recordType st
 	}
 
 	return prefix + DNSName[0] + suffix + "." + DNSName[1]
-}
-
-type kuadrantAffixNameMapper struct {
-	legacyMappers       map[string]nameMapper
-	prefix              string
-	wildcardReplacement string
-}
-
-type legacyMapperTemplate map[string]struct {
-	prefix              string
-	suffix              string
-	wildcardReplacement string
-}
-
-var _ nameMapper = kuadrantAffixNameMapper{}
-
-func newKuadrantAffixMapper(legacyMappersFor legacyMapperTemplate, prefix, wildcardReplacement string) nameMapper {
-	affixMappers := make(map[string]nameMapper)
-
-	for version, params := range legacyMappersFor {
-		affixMappers[version] = legacyAffixMappers[version](params.prefix, params.suffix, params.wildcardReplacement)
-	}
-	return kuadrantAffixNameMapper{
-		legacyMappers:       affixMappers,
-		prefix:              prefix,
-		wildcardReplacement: wildcardReplacement,
-	}
-}
-
-func (pr kuadrantAffixNameMapper) toEndpointName(txtDNSName, version string) (endpointName, recordType string) {
-	// legacy
-	if version != txtFormatVersion {
-		return pr.legacyMappers[version].toEndpointName(txtDNSName, version)
-	}
-
-	// ID-recordType-endpoint
-	dnsNameSplit := strings.SplitN(strings.TrimPrefix(strings.ToLower(txtDNSName), strings.ToLower(pr.prefix)), affixSeparator, 3)
-
-	// NPE guard
-	if len(dnsNameSplit) != 3 {
-		return "", ""
-	}
-
-	for _, rType := range getSupportedTypes() {
-		if dnsNameSplit[1] == strings.ToLower(rType) {
-			recordType = rType
-			break
-		}
-	}
-
-	endpointName = dnsNameSplit[2]
-
-	// undo wc replacement
-	if strings.HasPrefix(endpointName, pr.wildcardReplacement) {
-		strings.Replace(endpointName, pr.wildcardReplacement, "*", 1)
-	}
-
-	return endpointName, recordType
-}
-
-func (pr kuadrantAffixNameMapper) toTXTName(endpointDNSName, id, recordType string) string {
-	prefix := pr.prefix
-	if !strings.HasSuffix(prefix, affixSeparator) {
-		prefix = prefix + affixSeparator
-	}
-
-	dnsName := endpointDNSName
-
-	if strings.HasPrefix(dnsName, "*") {
-		dnsName = strings.Replace(dnsName, "*", pr.wildcardReplacement, 1)
-	}
-
-	return pr.prefix + hash.ToBase36HashLen(id, 8) + affixSeparator + strings.ToLower(recordType) + affixSeparator + dnsName
 }
